feat: add dark mode configuration tab
